# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Micro 01",
        "description": "API Micro 01",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get list of companies",
                "description": "Returns list of companies",
                "operationId": "getCompaniesList",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Store new company",
                "description": "Returns company data",
                "operationId": "storeCompanies",
                "parameters": [
                    {
                        "name": "category_id",
                        "description": "category_id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "description": "name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "description": "phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "whatsapp",
                        "description": "whatsapp",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "description": "email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "facebook",
                        "description": "facebook",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "instagram",
                        "description": "instagram",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "youtube",
                        "description": "youtube",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CompanyResource": {
                "title": "Companies",
                "description": "Companies model",
                "required": [],
                "properties": {
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id of Company",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Company",
                        "type": "string",
                        "example": "Coca-Cola"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of Company",
                        "type": "string",
                        "example": "99999-1992"
                    },
                    "whatsapp": {
                        "title": "whatsapp",
                        "description": "whatsapp of Company",
                        "type": "string",
                        "example": "99999-1992"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of Company",
                        "type": "string",
                        "example": "company@company.com"
                    },
                    "facebook": {
                        "title": "facebook",
                        "description": "facebook of Company",
                        "type": "string",
                        "example": "company profile"
                    },
                    "instagram": {
                        "title": "instagram",
                        "description": "instagram of Company",
                        "type": "string",
                        "example": "company profile"
                    },
                    "youtube": {
                        "title": "youtube",
                        "description": "youtube of Company",
                        "type": "string",
                        "example": "company channel"
                    }
                },
                "type": "object"
            }
        }
    }
}